+++ src/Markdown/FrontMatterLeagueCommonMarkExtension.php
@@ -13,6 +13,7 @@
 namespace Webuni\FrontMatter\Markdown;
 
 use Dflydev\DotAccessData\Data;
+use League\CommonMark\Environment\EnvironmentBuilderInterface;
 use League\CommonMark\Event\DocumentPreParsedEvent;
 use League\CommonMark\Extension\ExtensionInterface;
 use League\CommonMark\Input\MarkdownInput;
@@ -28,7 +29,7 @@ class FrontMatterLeagueCommonMarkExtension implements ExtensionInterface
         $this->frontMatter = $frontMatter;
     }
 
-    public function register($environment): void
+    public function register(EnvironmentBuilderInterface $environment): void
     {
         $environment->addEventListener(DocumentPreParsedEvent::class, [$this, 'parse']);
     }
@@ -39,12 +40,7 @@ class FrontMatterLeagueCommonMarkExtension implements ExtensionInterface
         $document = $this->frontMatter->parse($content);
         $data = $event->getDocument()->data;
 
-        if ($data instanceof Data) {
-            $data->import($document->getData(), Data::MERGE);
-        } else {
-            $event->getDocument()->data = array_merge($document->getData(), $data);
-        }
-
+        $data->import($document->getData(), Data::MERGE);
         $event->replaceMarkdown(new MarkdownInput($document->getContent()));
     }
 }