+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 * PHP 8.4 support
 * Drop PHP 7 support
 * Remove MtHAML filter
+* Add support for TOML v1.0.0
 
 2.0.0 (2024-02-22)
 ------------------